fix(jobs): make manual admission atomic
