+++ backend/jobs/cron_singleton_test.go
@@ -7,13 +7,10 @@ import (
 	"github.com/go-co-op/gocron/v2"
 )
 
-// TestSingletonModeBlocksOverlappingRunNow pins the guard against a manual
-// "Run now" and a cron tick processing the same queue files twice. Every
-// sched.NewJob call in this package passes
-// gocron.WithSingletonMode(gocron.LimitModeReschedule), and Job.RunNow() hands
-// the job to the executor over the same jobsIn channel a scheduled tick uses
-// (scheduler.go#selectRunJobRequest), so one singleton runner serializes both
-// the scheduled and the manual path.
+// TestSingletonModeBlocksOverlappingRunNow pins the scheduler-level backstop.
+// Every sched.NewJob call in this package passes
+// gocron.WithSingletonMode(gocron.LimitModeReschedule), so duplicate scheduler
+// submissions remain serialized in addition to the application-level runner.
 //
 // The first run is held hostage inside the task body, then a second RunNow is
 // fired. Under reschedule mode the second run is dropped and can never enter
@@ -91,40 +88,5 @@ func TestSingletonModeBlocksOverlappingRunNow(t *testing.T) {
 				t.Errorf("second run overlapped in-flight run = %v, want %v", gotOverlap, tc.wantOverlap)
 			}
 		})
-	}
-}
-
-func TestAutoDownloadRunGuardSpansJobGenerations(t *testing.T) {
-	entered := make(chan struct{})
-	release := make(chan struct{})
-	firstDone := make(chan bool, 1)
-
-	go func() {
-		firstDone <- runAutoDownloadIfIdle(func() {
-			close(entered)
-			<-release
-		})
-	}()
-
-	select {
-	case <-entered:
-	case <-time.After(5 * time.Second):
-		t.Fatal("first AutoDownload run never started")
-	}
-
-	secondEntered := false
-	if runAutoDownloadIfIdle(func() { secondEntered = true }) {
-		t.Error("replacement AutoDownload job ran while prior generation was active")
-	}
-	if secondEntered {
-		t.Error("replacement AutoDownload job entered guarded task")
-	}
-
-	close(release)
-	if !<-firstDone {
-		t.Error("first AutoDownload run did not acquire guard")
-	}
-	if !runAutoDownloadIfIdle(func() {}) {
-		t.Error("AutoDownload guard was not released after run completed")
 	}
 }
