+++ backend/jobs/download_auto.go
@@ -6,23 +6,10 @@ import (
 	"aura/logging"
 	"context"
 	"runtime/debug"
-	"sync"
 
 	"github.com/go-co-op/gocron/v2"
 	"github.com/google/uuid"
 )
-
-// Spans scheduler job IDs so config reschedules cannot overlap an in-flight run.
-var autodownloadRunMu sync.Mutex
-
-func runAutoDownloadIfIdle(run func()) bool {
-	if !autodownloadRunMu.TryLock() {
-		return false
-	}
-	defer autodownloadRunMu.Unlock()
-	run()
-	return true
-}
 
 func StartAutoDownloadJob() error {
 	mu.Lock()
@@ -54,7 +41,7 @@ func StartAutoDownloadJob() error {
 
 	job, err := sched.NewJob(
 		gocron.CronJob(spec, false),
-		gocron.NewTask(func() {
+		gocron.NewTask(configureJobRunner("AutoDownload Job", func() {
 			defer func() {
 				if r := recover(); r != nil {
 					logging.LOGGER.Error().
@@ -64,24 +51,20 @@ func StartAutoDownloadJob() error {
 						Msg("PANIC: in scheduled AutoDownload Job")
 				}
 			}()
-			if !runAutoDownloadIfIdle(func() {
-				ctx, ld := logging.CreateLoggingContext(context.Background(), "Cron Job")
-				action := ld.AddAction("AutoDownload Check", logging.LevelInfo)
-				ctx = logging.WithCurrentAction(ctx, action)
-				Err := autodownload.CheckAllItems(ctx)
-				if Err.Message != "" {
-					logging.LOGGER.Error().Timestamp().Str("error", Err.Message).
-						Str("next_run", formatNextRun(autodownloadJob)).
-						Msg("Error running AutoDownload Job")
-				} else {
-					logging.LOGGER.Info().Timestamp().
-						Str("next_run", formatNextRun(autodownloadJob)).
-						Msg("AutoDownload Job Completed")
-				}
-			}) {
-				logging.LOGGER.Warn().Timestamp().Msg("AutoDownload Job skipped because another run is in progress")
+			ctx, ld := logging.CreateLoggingContext(context.Background(), "Cron Job")
+			action := ld.AddAction("AutoDownload Check", logging.LevelInfo)
+			ctx = logging.WithCurrentAction(ctx, action)
+			Err := autodownload.CheckAllItems(ctx)
+			if Err.Message != "" {
+				logging.LOGGER.Error().Timestamp().Str("error", Err.Message).
+					Str("next_run", formatNextRun(autodownloadJob)).
+					Msg("Error running AutoDownload Job")
+			} else {
+				logging.LOGGER.Info().Timestamp().
+					Str("next_run", formatNextRun(autodownloadJob)).
+					Msg("AutoDownload Job Completed")
 			}
-		}),
+		}).runScheduled),
 		gocron.WithName("AutoDownload Job"),
 		gocron.WithSingletonMode(gocron.LimitModeReschedule),
 	)
