+++ backend/jobs/kometa_import.go
@@ -40,7 +40,7 @@ func StartKometaImportJob() error {
 
 	job, err := sched.NewJob(
 		gocron.CronJob(spec, false),
-		gocron.NewTask(func() {
+		gocron.NewTask(configureJobRunner("Kometa Asset Import Job", func() {
 			defer func() {
 				if r := recover(); r != nil {
 					logging.LOGGER.Error().
@@ -53,7 +53,7 @@ func StartKometaImportJob() error {
 			if started := kometa.StartImport(); !started {
 				logging.LOGGER.Warn().Timestamp().Msg("Kometa Asset Import skipped (already running or not enabled)")
 			}
-		}),
+		}).runScheduled),
 		gocron.WithName("Kometa Asset Import Job"),
 		gocron.WithSingletonMode(gocron.LimitModeReschedule),
 	)
