+++ backend/jobs/mediux_check_link.go
@@ -29,7 +29,7 @@ func StartCheckMediuxSiteLinkJob() error {
 	spec := "*/60 * * * *"
 	job, err := sched.NewJob(
 		gocron.CronJob(spec, false),
-		gocron.NewTask(func() {
+		gocron.NewTask(configureJobRunner("Check Mediux Site Link Availability Job", func() {
 			defer func() {
 				if r := recover(); r != nil {
 					logging.LOGGER.Error().Timestamp().Interface("recover", r).Msg("PANIC: in scheduled CheckMediuxSiteLinkJob")
@@ -40,7 +40,7 @@ func StartCheckMediuxSiteLinkJob() error {
 			ctx = logging.WithCurrentAction(ctx, action)
 			mediux.CheckSiteLinkAvailability()
 			ld.Log()
-		}),
+		}).runScheduled),
 		gocron.WithName("Check Mediux Site Link Availability Job"),
 		gocron.WithSingletonMode(gocron.LimitModeReschedule),
 	)
